AP_SpdHgtControl: use float for get_throttle_demand

+++ libraries/AP_SpdHgtControl/AP_SpdHgtControl.h
@@ -35,7 +35,7 @@ public:
 
 	// demanded throttle in percentage
 	// should return 0 to 100
-	virtual int32_t get_throttle_demand(void)=0;
+	virtual float get_throttle_demand(void)=0;
 	
 	// demanded pitch angle in centi-degrees
 	// should return -9000 to +9000
